Directory Mistake Change

+++ GameEngine Practice/Unity Clone Coding/Engine_Source/Source/PEInput.cpp
@@ -102,11 +102,11 @@ namespace PracticeEngine
 		mMousePosition.x = -1.0f;
 		mMousePosition.y = -1.0f;
 
-		if ((UINT)mousePos.x > 0 && (UINT)mousePos.x < width)
-			mMousePosition.x = (float)mousePos.x;
+		if (mousePos.x > 0 && mousePos.x < width)
+			mMousePosition.x = mousePos.x;
 
-		if ((UINT)mousePos.y > 0 && (UINT)mousePos.y < height)
-			mMousePosition.y = (float)mousePos.y;
+		if (mousePos.y > 0 && mousePos.y < height)
+			mMousePosition.y = mousePos.y;
 	}
 	void Input::clearKeys()
 	{
